core/game/requirements.txt
asgiref==3.8.1
channels==4.2.2
channels_redis==4.2.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
redis==5.2.1
sqlparse==0.5.3
Pillow==9.5.0
wheel==0.36.0
httpx==0.28.1